requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
redis
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
pydantic-settings
apscheduler
pytest
pytest-cov
locust
fakeredis
